[clang-tidy] Fix msgid type of TimerImpl_privdata

+++ src/TimerImpl_privdata.h
@@ -4,7 +4,7 @@
 #include <tgbot/tgbot.h>
 
 struct TimerImpl_privdata {
-    int64_t messageid;
+    int32_t messageid;
     const TgBot::Bot& bot;
     bool botcanpin, sendendmsg;
     int64_t chatid;
